Conform to CrateLookup trait for CrateApi
