+++ src/crates/mod.rs
@@ -1,3 +1,4 @@
+pub mod api;
 pub mod cache;
 pub mod sparse;
